% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_flux_columns}
\alias{add_flux_columns}
\title{Add NA flux columns, if necessary}
\usage{
add_flux_columns(dsd)
}
\arguments{
\item{dsd}{Dataset data, a \code{data.frame}}
}
\value{
The data frame with both CSR_FLUX_CH4 and CSR_FLUX_CO2.
}
\description{
Add NA flux columns, if necessary
}
\keyword{internal}
