% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_port_column}
\alias{add_port_column}
\title{Add port column, if necessary, and convert to numeric.}
\usage{
add_port_column(dsd)
}
\arguments{
\item{dsd}{Dataset data, a \code{data.frame}}
}
\value{
The data frame with a numeric PORT column.
}
\description{
Add port column, if necessary, and convert to numeric.
}
\keyword{internal}
