% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_timestamps}
\alias{calc_timestamps}
\title{Calculate all timestamps based on timestamps given and measurement length}
\usage{
calc_timestamps(dsd, ml, tf, tz)
}
\arguments{
\item{dsd}{Dataset data, a data frame}

\item{ml}{Measurement length, numeric (seconds)}

\item{tf}{Timestamp format, a \code{\link{strptime}} format string}

\item{tz}{Timezone, character}
}
\value{
A list with (i) the new dataset data.frame, (ii) a logical vector
indicating which ones are invalid (and thus now NA), and
(iii) a string of examples of bad timestamps.
}
\description{
Calculate all timestamps based on timestamps given and measurement length
}
\note{
This is called from \code{\link{read_raw_dataset}}.
}
