% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{check_file_description_consistency}
\alias{check_file_description_consistency}
\title{Check that files have descriptions and descriptions have files.}
\usage{
check_file_description_consistency(all_files, file_descriptions)
}
\arguments{
\item{all_files}{Vector of filenames, characfter}

\item{file_descriptions}{Vector of file descriptions, character}
}
\value{
Nothing; run for side effect of error if necessary.
}
\description{
Check that files have descriptions and descriptions have files.
}
\note{
Called only by \code{\link{csr_make_release}}.
}
\keyword{internal}
