% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_data}
\alias{combine_data}
\title{Combine all datasets into a single list}
\usage{
combine_data(datasets, ...)
}
\arguments{
\item{datasets}{Character vector of dataset names}

\item{...}{Dataset objects}
}
\value{
A single list with all datasets
}
\description{
Combine all datasets into a single list
}
\keyword{internal}
