% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{combined_report}
\alias{combined_report}
\title{Run all individual dataset reports}
\usage{
combined_report(all_data, output_dir = "~/Desktop/Reports/")
}
\arguments{
\item{all_data}{A list of \code{cosore} datasets.}

\item{output_dir}{Output directory}
}
\value{
Nothing; run for its side effects.
}
\description{
Run all individual dataset reports
}
