% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_interval}
\alias{compute_interval}
\title{Compute measurement interval for a dataset}
\usage{
compute_interval(dsd)
}
\arguments{
\item{dsd}{Dataset data (a data frame)}
}
\value{
Median interval between timestamps.
}
\description{
Compute measurement interval for a dataset
}
\note{
This is used by the reports.
}
