% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_ancillary_timestamps}
\alias{convert_ancillary_timestamps}
\title{Convert ancillary timestamps to POSIXct}
\usage{
convert_ancillary_timestamps(ancillary, description)
}
\arguments{
\item{ancillary}{An ancillary data frame}

\item{description}{A description data frame}
}
\value{
The ancillary data frame
}
\description{
Convert ancillary timestamps to POSIXct
}
\note{
Called only by \code{\link{read_dataset}}.
}
\keyword{internal}
