% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_and_qc_timestamp}
\alias{convert_and_qc_timestamp}
\title{Convert a vector of strings to POSIXct timestamps.}
\usage{
convert_and_qc_timestamp(ts, timestamp_format, time_zone)
}
\arguments{
\item{ts}{Timestamps, character}

\item{timestamp_format}{Formatting string that \code{\link{as.POSIXct}} accepts.
This normally comes from \code{CSR_TIMESTAMP_FORMAT} in the DESCRIPTION file.}

\item{time_zone}{Time zone that \code{\link{as.POSIXct}} accepts.
This normally comes from \code{CSR_TIMESTAMP_TZ} in the DESCRIPTION file.}
}
\value{
A list with (i) a vector of \code{POSIXct} timestamps, (ii) a logical vector
indicating which ones are invalid (and thus now NA), and (iii) a string of examples of
bad timestamps.
}
\description{
Convert a vector of strings to POSIXct timestamps.
}
\keyword{internal}
