% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_numeric}
\alias{convert_to_numeric}
\title{Convert a vector to numeric, giving an informative warning.}
\usage{
convert_to_numeric(x, name, warn = TRUE)
}
\arguments{
\item{x}{Vector}

\item{name}{Name of object (for warning)}

\item{warn}{Warn if non-numeric? Logical}
}
\value{
A numeric vector.
}
\description{
Convert a vector to numeric, giving an informative warning.
}
\keyword{internal}
