% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{csr_build}
\alias{csr_build}
\title{Build the COSORE dataset}
\usage{
csr_build(raw_data, return_all_data = TRUE)
}
\arguments{
\item{raw_data}{The raw data folder to use, character path}

\item{return_all_data}{Return all data generated (invisibly)?, Logical}
}
\value{
All the built data, invisibly.
}
\description{
Build the COSORE dataset
}
\examples{
\dontrun{
all_data <- csr_build("../rawdata/")  # build with raw data
run_all_reports(all_data)
}
}
