% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{csr_data}
\alias{csr_data}
\title{Extract all the \code{data} tables from the COSORE dataset.}
\usage{
csr_data(all)
}
\arguments{
\item{all}{All data, a list of lists}
}
\value{
A \code{\link{data.frame}}.
}
\description{
Extract all the \code{data} tables from the COSORE dataset.
}
