% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{csr_database}
\alias{csr_database}
\title{Return an overview of available datasets}
\usage{
csr_database(all)
}
\arguments{
\item{all}{All data, a list of lists}
}
\value{
A `data.frame` or `tibble` with data about the constitutent datasets of COSORE.
}
\description{
Return an overview of available datasets
}
\note{
This is a convenience function, as it simply calls
\code{\link{csr_table}} to return the joined \code{description} tables.
}
