% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{csr_dataset}
\alias{csr_dataset}
\title{Return an individual dataset}
\usage{
csr_dataset(dataset)
}
\arguments{
\item{dataset}{Name of dataset, character}
}
\value{
A list with (at least) elements:
\item{description}{Contents of \code{DESCRIPTION.txt} file}
\item{contributors}{Contents of \code{CONTRIBUTORS.txt} file}
\item{ports}{Contents of \code{PORTS.txt} file}
\item{data}{Continuous soil respiration data, parsed into a \code{data.frame}}
\item{diagnostics}{Diagnostics on the data parsing and QC process}
\item{ancillary}{Ancillary site information}
}
\description{
Return an individual dataset
}
