% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{csr_element}
\alias{csr_element}
\title{Extract one part (\code{description}, \code{data}, etc.) of all datasets.}
\usage{
csr_element(all, element)
}
\arguments{
\item{all}{All data, a list of lists}

\item{element}{Name of table to extract, character}
}
\value{
A \code{\link{data.frame}}.
}
\description{
Extract one part (\code{description}, \code{data}, etc.) of all datasets.
}
