% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{csr_emails}
\alias{csr_emails}
\title{Extract contact emails}
\usage{
csr_emails(all)
}
\arguments{
\item{all}{All data, a list of lists}
}
\value{
A pasted string of emails, separated by semicolons.
}
\description{
Extract contact emails
}
