% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{csr_make_release}
\alias{csr_make_release}
\title{Make a new COSORE release based on current package datasets.}
\usage{
csr_make_release(
  path,
  vignette_rebuilt = FALSE,
  force = FALSE,
  run_report = TRUE,
  zip_release = TRUE,
  datasets = list_datasets(),
  sys_call = system2
)
}
\arguments{
\item{path}{Path to write files to; must already exist}

\item{vignette_rebuilt}{Has vignette been rebuilt? Logical}

\item{force}{Ignore git dirty status? Logical}

\item{run_report}{Include \code{combined_report.Rmd} in release? For testing}

\item{zip_release}{Run /code{utils::zip()} on output files? For testing}

\item{datasets}{Datasets to include, character; used for testing}

\item{sys_call}{System call to use, function; testing only; normally
\code{\link{system2}}}
}
\value{
Fully qualified name of zip file containing release.
}
\description{
Make a new COSORE release based on current package datasets.
}
\details{
To make a new release, the git working directory
must be clean (unless \code{force = TRUE}), a \code{path} must
be specified, and there should be file descriptors (in the code)
for all files that will be zipped into the release;
this information is inserted into the \code{README.md} file.
Dynamic information like the current date, git commit hash,
database version, and
database size are all copied into \code{README.md} as well.
}
\note{
If a \code{CSR_EMBARGO} field exists in the DESCRIPTION
file, no data will be released from that dataset.
}
