% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{csr_metadata}
\alias{csr_metadata}
\title{Return metadata information about all COSORE tables}
\usage{
csr_metadata()
}
\value{
Metadata information (as a data frame) about all COSORE tables.
}
\description{
Return metadata information about all COSORE tables
}
