% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{csr_remove_stan_data}
\alias{csr_remove_stan_data}
\title{Remove standardized dataset(s)}
\usage{
csr_remove_stan_data(path, datasets = list_datasets())
}
\arguments{
\item{path}{Path, typically \code{inst/extdata/datasets/}}

\item{datasets}{Optional character vector of datasets. If not
supplied, the output of \code{\link{list_datasets}} is used.}
}
\value{
Nothing.
}
\description{
Utility function that remove standardized datasets from a
directory tree, specifically \code{path/dataset/data/*.RDS}.
}
