% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{csr_report_database}
\alias{csr_report_database}
\title{Run a summary report on the entire COSORE database}
\usage{
csr_report_database(output_dir = "~/Desktop/Reports/")
}
\arguments{
\item{output_dir}{Output directory}
}
\value{
Nothing; run for its side effects.
}
\description{
Run a summary report on the entire COSORE database
}
