% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{csr_report_dataset}
\alias{csr_report_dataset}
\title{Run a single dataset report}
\usage{
csr_report_dataset(
  ds,
  output_dir = "~/Desktop/Reports/",
  quiet = FALSE,
  quick = FALSE
)
}
\arguments{
\item{ds}{The dataset for which to generate a report.}

\item{output_dir}{Output directory, character}

\item{quiet}{Passed on to \code{rmarkdown::render}}

\item{quick}{Passed on to \code{rmarkdown::render}}
}
\value{
Nothing; run for its side effects.
}
\description{
Run a single dataset report, writing the resulting \code{html}
file into a specified directory.
}
