% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{csr_reports_all}
\alias{csr_reports_all}
\title{Run all individual dataset and database reports}
\usage{
csr_reports_all()
}
\value{
Nothing; run for its side effects.
}
\description{
Run all individual dataset and database reports
}
