% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_dir}
\alias{data_dir}
\title{Find the directory holding standardized data for a dataset}
\usage{
data_dir(dataset)
}
\arguments{
\item{dataset}{Name of dataset, character}
}
\value{
Path to the standardized data directory of a dataset
}
\description{
Find the directory holding standardized data for a dataset
}
\keyword{internal}
