% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{embargoed}
\alias{embargoed}
\title{Check whether a dataset is embargoed}
\usage{
embargoed(ds)
}
\arguments{
\item{ds}{The dataset}
}
\value{
TRUE or FALSE.
}
\description{
Check whether a dataset is embargoed
}
\keyword{internal}
