% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{extract_line}
\alias{extract_line}
\title{Parse labels from information files}
\usage{
extract_line(file_data, line_label, required = TRUE, sep = ":",
  numeric_data = FALSE)
}
\arguments{
\item{file_data}{Character vector of file lines}

\item{line_label}{Character label}

\item{sep}{Separator}

\item{numeric_data}{Numeric data? Logical}
}
\value{
Value (character) of data opposite label
}
\description{
Parse labels from information files
}
\examples{
extract_line(c("1:one", "2:two"), "two")
extract_line(c("one:1", "two:2"), "two", numeric_data = TRUE)
}
\keyword{internal}
