% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{extract_table}
\alias{extract_table}
\title{Internal function to extract a table from a dataset.}
\usage{
extract_table(dataset_name, table, quiet, retrieval_function = csr_dataset)
}
\arguments{
\item{dataset_name}{Dataset name, character}

\item{table}{Table name, character}

\item{quiet}{Print progress messages and warnings? Logical}

\item{retrieval_function}{Dataset retrieval function,
normally \code{\link{csr_dataset}}; used only in testing}
}
\value{
The extracted data frame.
}
\description{
Internal function to extract a table from a dataset.
}
\keyword{internal}
