% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fractional_doy}
\alias{fractional_doy}
\title{fractional_doy}
\usage{
fractional_doy(year, doy)
}
\arguments{
\item{year}{Year, integer}

\item{doy}{Day of year, numeric (xxx.xxx)}
}
\value{
A string giving day of year and time in \code{\%j \%T} format; see \code{\link{strptime}}.
}
\description{
Convert a day of year (DOY) given as xxx.xxx to a day of year + time of day.
This can vary due to leap years, so we need the year in question as well.
}
\keyword{internal}
