% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gases_string}
\alias{gases_string}
\title{Make a string (for diagnostics) saying which gases are represented in a dataset}
\usage{
gases_string(dsd)
}
\arguments{
\item{dsd}{Dataset data, a data.frame}
}
\value{
String with gases separated by commas
}
\description{
Make a string (for diagnostics) saying which gases are represented in a dataset
}
\keyword{internal}
