% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_defaults}
\alias{insert_defaults}
\title{Insert default fields into a data frame}
\usage{
insert_defaults(d, extra_fields)
}
\arguments{
\item{d}{A tibble or data.frame}

\item{extra_fields}{A list of fields (the names) and their default values (elements)}
}
\value{
The (possibly modified) data frame.
}
\description{
Insert default fields into a data frame
}
\keyword{internal}
