% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insert_line}
\alias{insert_line}
\title{Insert new line(s) into existing metadata files.}
\usage{
insert_line(file, pattern, newlines, path = "inst/extdata",
  write_files = TRUE)
}
\arguments{
\item{file}{Filename, character}

\item{pattern}{Pattern to search for, a regular expression}

\item{newlines}{New line(s) to write, character vector}

\item{path}{Path to search; normally the metadata are in \code{inst/extdata}}

\item{write_files}{Write resulting files back out? Logical}
}
\value{
Nothing; run for side effects
}
\description{
Insert new line(s) into existing metadata files.
}
