% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{list_datasets}
\alias{list_datasets}
\title{List available datasets}
\usage{
list_datasets(path = resolve_dataset(""))
}
\arguments{
\item{path}{Path name}
}
\value{
Character vector of datasets.
}
\description{
List available datasets
}
\examples{
list_datasets()
}
