% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{make_cosore_release}
\alias{make_cosore_release}
\title{Make a new COSORE release}
\usage{
make_cosore_release(all_data, dest)
}
\arguments{
\item{all_data}{A list of \code{cosore} datasets.}

\item{dest}{Dest}
}
\value{
Nothing; run for its side effects.
}
\description{
Make a new COSORE release
}
