% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{map_columns}
\alias{map_columns}
\title{Map data columns to new names/values.}
\usage{
map_columns(dat, columns)
}
\arguments{
\item{dat}{Dataset data, a \code{data.frame}}

\item{columns}{Column mapping data from the \code{COLUMNS.txt} file, a \code{data.frame}}
}
\value{
The \code{dat} data frame with column names transformed, and possibly
computed, as defined by \code{columns}.
}
\description{
Map data columns to new names/values.
}
\examples{
dat <- data.frame(x = 1:3)
columns <- data.frame(Database = "y", Dataset = "x", Computation = "x * 2")
map_columns(dat, columns)  # produces a data.frame(y = c(2, 4, 6))
}
\keyword{internal}
