% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{minigather}
\alias{minigather}
\title{Simple function to gather (reshape) data.}
\usage{
minigather(
  x,
  gather_cols,
  category_col_name,
  value_col_name,
  new_categories = NULL
)
}
\arguments{
\item{x}{Data frame to reshape}

\item{gather_cols}{Names of columns to gather, character vector}

\item{category_col_name}{Name of resulting category column, character}

\item{value_col_name}{Name of resulting value column, character}

\item{new_categories}{Optional new categories to use
in place of gathered column names, character vector}
}
\value{
The data frame in long format.
}
\description{
Simple function to gather (reshape) data.
}
\note{
Base R's \code{\link{reshape}} sucks, while \code{\link{tidyr::gather}}
brings in \code{tidyr}, which imports \code{dplyr}, ...way too heavy.
}
\examples{
cosore:::minigather(cars, c("speed", "dist"), "var", "val")
}
\keyword{internal}
