% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_d20190415_HF068}
\alias{parse_d20190415_HF068}
\title{Parse a file from Harvard Forest HF068 dataset.}
\usage{
parse_d20190415_HF068(path, UTC_offset)
}
\arguments{
\item{path}{Data directory path, character}

\item{UTC_offset}{Offset from UTC in hours, numeric}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a file from Harvard Forest HF068 dataset.
}
