% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_d20190430_DESAI}
\alias{parse_d20190430_DESAI}
\title{Parse a custom eofFD (forced diffusion) file from d20190430_DESAI.}
\usage{
parse_d20190430_DESAI(path, UTC_offset)
}
\arguments{
\item{path}{Data directory path, character}

\item{UTC_offset}{Offset from UTC in hours, numeric}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a custom eofFD (forced diffusion) file from d20190430_DESAI.
}
