% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_d20190527_GOULDEN}
\alias{parse_d20190527_GOULDEN}
\title{Parse a custom file from d20190527_GOULDEN}
\usage{
parse_d20190527_GOULDEN(path)
}
\arguments{
\item{path}{Data directory path, character}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a custom file from d20190527_GOULDEN
}
\keyword{internal}
