% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_d20190617_SCOTT_SRM}
\alias{parse_d20190617_SCOTT_SRM}
\title{Parse a custom file from d20190617_SCOTT_SRM}
\usage{
parse_d20190617_SCOTT_SRM(path)
}
\arguments{
\item{path}{Data directory path, character}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a custom file from d20190617_SCOTT_SRM
}
\keyword{internal}
