% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_d20190617_SCOTT_WHS}
\alias{parse_d20190617_SCOTT_WHS}
\title{Parse a custom file from d20190617_SCOTT_WHS}
\usage{
parse_d20190617_SCOTT_WHS(path)
}
\arguments{
\item{path}{Data directory path, character}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a custom file from d20190617_SCOTT_WHS
}
\keyword{internal}
