% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_d20190617_SCOTT_xxx}
\alias{parse_d20190617_SCOTT_xxx}
\title{Parse one of the d20190617_SCOTT custom files}
\usage{
parse_d20190617_SCOTT_xxx(path, skip, ports)
}
\arguments{
\item{path}{Ports, an integer vector}

\item{skip}{Lines to skip, integer}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse one of the d20190617_SCOTT custom files
}
\keyword{internal}
