% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_d20190830_LIANG}
\alias{parse_d20190830_LIANG}
\title{Parse a custom file from d20190830_LIANG}
\usage{
parse_d20190830_LIANG(path)
}
\arguments{
\item{path}{Data directory path, character}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a custom file from d20190830_LIANG
}
\note{
This is a complicated one: the authors posted their data on
Figshare (good), but provide only raw CO2 concentration data (good
but a pain). Need to read each file, compute the flux for each
chamber at each timestep, and stitch together. This is slow.
}
\keyword{internal}
