% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_d20200109_HIRANO_PDF}
\alias{parse_d20200109_HIRANO_PDF}
\title{Parse a custom file from d20200109_HIRANO_PDF.}
\usage{
parse_d20200109_HIRANO_PDF(path)
}
\arguments{
\item{path}{Data directory path, character}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a custom file from d20200109_HIRANO_PDF.
}
\keyword{internal}
