% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_d20200122_BLACK}
\alias{parse_d20200122_BLACK}
\title{Parse a custom file from d20200108_JASSAL.}
\usage{
parse_d20200122_BLACK(path)
}
\arguments{
\item{path}{Data directory path, character}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a custom file from d20200108_JASSAL.
}
\keyword{internal}
