% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_d20200214_BALDOCCHI_TONZI}
\alias{parse_d20200214_BALDOCCHI_TONZI}
\title{Parse a custom file from d20200214_BALDOCCHI_TONZI.}
\usage{
parse_d20200214_BALDOCCHI_TONZI(path)
}
\arguments{
\item{path}{Data directory path, character}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a custom file from d20200214_BALDOCCHI_TONZI.
}
\keyword{internal}
