% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_d20200328_UEYAMA_YAMASHIRO}
\alias{parse_d20200328_UEYAMA_YAMASHIRO}
\title{Parse a custom file from d20200328_UEYAMA_YAMASHIRO}
\usage{
parse_d20200328_UEYAMA_YAMASHIRO(path)
}
\arguments{
\item{path}{Data directory path, character}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a custom file from d20200328_UEYAMA_YAMASHIRO
}
\keyword{internal}
