% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_d20200407_WANG}
\alias{parse_d20200407_WANG}
\title{Parse a custom file from parse_d20200407_WANG}
\usage{
parse_d20200407_WANG(path)
}
\arguments{
\item{path}{Data directory path, character}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a custom file from parse_d20200407_WANG
}
\keyword{internal}
