% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_d20200419_PEREZ-QUEZADA}
\alias{parse_d20200419_PEREZ-QUEZADA}
\title{Parse a custom file from d20200419_PEREZ_QUEZADA}
\usage{
`parse_d20200419_PEREZ-QUEZADA`(path)
}
\arguments{
\item{path}{Data directory path, character}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a custom file from d20200419_PEREZ_QUEZADA
}
\keyword{internal}
