% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_d20200423_SANCHEZ-CANETE}
\alias{parse_d20200423_SANCHEZ-CANETE}
\title{Parse a custom file from d20200423_SANCHEZ-CANETE}
\usage{
`parse_d20200423_SANCHEZ-CANETE`(path)
}
\arguments{
\item{path}{Data directory path, character}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a custom file from d20200423_SANCHEZ-CANETE
}
\keyword{internal}
