% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_eosFD}
\alias{parse_eosFD}
\title{Parse a eofFD (forced diffusion) file.}
\usage{
parse_eosFD(path, UTC_offset)
}
\arguments{
\item{path}{Data directory path, character}

\item{UTC_offset}{Offset from UTC in hours, numeric}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a eofFD (forced diffusion) file.
}
