% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_IRGA_ydh}
\alias{parse_IRGA_ydh}
\title{Parse a file with timestamp in year-doy-hour format.}
\usage{
parse_IRGA_ydh(path, UTC_offset)
}
\arguments{
\item{path}{Data directory path, character}

\item{UTC_offset}{Offset from UTC in hours, numeric}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a file with timestamp in year-doy-hour format.
}
