% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_LI-6252_d20190504_DAVIDSON_hf006-03}
\alias{parse_LI-6252_d20190504_DAVIDSON_hf006-03}
\title{Parse a custom file from d20190504_DAVIDSON_hf006-03.}
\usage{
"parse_LI-6252_d20190504_DAVIDSON_hf006-03"(path, UTC_offset)
}
\arguments{
\item{path}{Data directory path, character}

\item{UTC_offset}{Offset from UTC in hours, numeric}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a custom file from d20190504_DAVIDSON_hf006-03.
}
