% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_LI-8100A_d20190517_MAURITZ}
\alias{parse_LI-8100A_d20190517_MAURITZ}
\title{Parse a file from d20190517_MAURITZ dataset--fractional hours in separate column.}
\usage{
"parse_LI-8100A_d20190517_MAURITZ"(path)
}
\arguments{
\item{path}{Data directory path, character}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a file from d20190517_MAURITZ dataset--fractional hours in separate column.
}
