% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-LI8100.R
\name{parse_LI-8100A_PROCESSED}
\alias{parse_LI-8100A_PROCESSED}
\title{Loop through directory and read processed multiplexed Licor-8100 data}
\usage{
"parse_LI-8100A_PROCESSED"(path)
}
\arguments{
\item{path}{Data directory path, character}
}
\value{
A data frame with all data read from file(s).
}
\description{
Loop through directory and read processed multiplexed Licor-8100 data
}
\note{
Processed (in the Licor application) data consists of a tab-delimited
text file with a standard set of columns.
}
