% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-LI8100.R
\name{parse_LI-8100A_RAW}
\alias{parse_LI-8100A_RAW}
\title{Read raw multiplexed Licor-8100 data}
\usage{
"parse_LI-8100A_RAW"(path)
}
\arguments{
\item{path}{Data directory path, character}
}
\value{
A data frame with all data read from file(s).
}
\description{
Read raw multiplexed Licor-8100 data
}
\keyword{internal}
