% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_LI-820_PROCESSED_CSV}
\alias{parse_LI-820_PROCESSED_CSV}
\title{Parse a file from Harvard Forest HF068 dataset.}
\usage{
"parse_LI-820_PROCESSED_CSV"(path)
}
\arguments{
\item{path}{Data directory path, character}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a file from Harvard Forest HF068 dataset.
}
