% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-LI8100.R
\name{parse_LI8100_file}
\alias{parse_LI8100_file}
\title{Parse a LI-8100 (with LI-8150 multiplexer) data file.}
\usage{
parse_LI8100_file(filename, UTC_offset)
}
\arguments{
\item{filename}{Filename, character}

\item{UTC_offset}{Offset from UTC in hour, numeric}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a LI-8100 (with LI-8150 multiplexer) data file.
}
