% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-LI8100.R
\name{parse_LI8100_processed}
\alias{parse_LI8100_processed}
\title{Loop through directory and read processed multiplexed Licor-8100 data}
\usage{
parse_LI8100_processed(path, UTC_offset)
}
\arguments{
\item{path}{Directory path, character}

\item{UTC_offset}{Offset from UTC in hour, numeric}

\item{port_data}{Port data, a list returned by \code{\link{read_ports_file}}}
}
\value{
A data frame with all data read from file(s).
}
\description{
Loop through directory and read processed multiplexed Licor-8100 data
}
\note{
Processed (in the Licor application) data consists of a tab-delimited
text file with a standard set of columns.
}
