% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-LI8100.R
\name{parse_LI8100_record}
\alias{parse_LI8100_record}
\title{Parse a LI-8100 (with LI-8150 multiplexer) data record}
\usage{
parse_LI8100_record(record, results, i, bfn, rec_start)
}
\arguments{
\item{record}{Character vector containing record data extracted from file}

\item{results}{Pre-allocated results, a data.frame}

\item{i}{Position in \code{results} in which to insert extracted data}

\item{rec_start}{Record position within file, for error reporting, numeric}

\item{brn}{Base filename being used, for error reporting, character}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a LI-8100 (with LI-8150 multiplexer) data record
}
\note{
This is called only by \code{\link{parse_LI8100_file}}.
}
\keyword{internal}
