% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-LI8100.R
\name{parse_PROCESSED}
\alias{parse_PROCESSED}
\title{Read processed data in whitespace-delimited format}
\usage{
parse_PROCESSED(path)
}
\arguments{
\item{path}{Data directory path, character}
}
\value{
A data frame with all data read from file(s).
}
\description{
Read processed data in whitespace-delimited format
}
\note{
Processed (in the Licor application) data consists of a tab-delimited
text file with a standard set of columns.
}
\keyword{internal}
