% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-processed.R
\name{parse_PROCESSED_CSV}
\alias{parse_PROCESSED_CSV}
\title{Read processed data in CSV format}
\usage{
parse_PROCESSED_CSV(path)
}
\arguments{
\item{path}{Data directory path, character}
}
\value{
A data frame with all data read from file(s).
}
\description{
Read processed data in CSV format
}
\note{
Processed (in the Licor application) data consists of a tab-delimited
text file with a standard set of columns.
}
\keyword{internal}
