% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-others.R
\name{parse_YEARDOYHOUR}
\alias{parse_YEARDOYHOUR}
\title{Parse a file with timestamp in year-doy-hour format.}
\usage{
parse_YEARDOYHOUR(path)
}
\arguments{
\item{path}{Data directory path, character}
}
\value{
A \code{data.frame} containing extracted data.
}
\description{
Parse a file with timestamp in year-doy-hour format.
}
