% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{qaqc_data}
\alias{qaqc_data}
\title{QA/QC a dataset.}
\usage{
qaqc_data(dsd, diag, remove_na = TRUE, remove_error = TRUE)
}
\arguments{
\item{dsd}{Dataset data, a tibble}

\item{diag}{Diagnostics data, a tibble}

\item{remove_na}{Remove NA flux values? Logical}

\item{remove_error}{Remove records marked as errors? Logical}
}
\value{
A list with the `dsd` and `diag` tibbles.
}
\description{
QA/QC a dataset.
}
\note{
This is normally called from \code{\link{read_raw_dataset}}.
}
\keyword{internal}
