% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbind_all}
\alias{rbind_all}
\title{Bind a list of data frames (with possibly non-identical column names) together.}
\usage{
rbind_all(x)
}
\arguments{
\item{x}{List of data frames}
}
\value{
A single \code{data.frame} with all data together
}
\description{
Bind a list of data frames (with possibly non-identical column names) together.
}
\keyword{internal}
