% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_ancillary_file}
\alias{read_ancillary_file}
\title{Read dataset ANCILLARY file}
\usage{
read_ancillary_file(dataset_name, file_data = NULL)
}
\arguments{
\item{dataset_name}{Dataset name, character}

\item{file_data}{File data, character vector; optional for testing}
}
\value{
A \code{data.frame} containing any data in the file.
}
\description{
Read dataset ANCILLARY file
}
\note{
This is simply a comma-separated table.
}
\keyword{internal}
