% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_columns_file}
\alias{read_columns_file}
\title{Read dataset COLUMNS file}
\usage{
read_columns_file(dataset_name, file_data = NULL)
}
\arguments{
\item{dataset_name}{Dataset name, character}

\item{file_data}{File data, character vector; optional for testing}
}
\value{
A \code{data.frame} with the following columns:
\item{Database}{Database column name, character}
\item{Dataset}{Dataset column name, character}
\item{Computation}{Optional computation R-parseable to perform, character}
\item{Port}{Optional port number, integer}
\item{Notes}{Optional notes, character}
}
\description{
Read dataset COLUMNS file
}
\keyword{internal}
