% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_csv_data}
\alias{read_csv_data}
\title{Read comma-separated data from a character vector}
\usage{
read_csv_data(file_data, required = NULL)
}
\arguments{
\item{file_data}{File data to read, character vector}

\item{required}{Vector of column names that must be all filled in, optional}
}
\value{
A data frame with loaded data.
}
\description{
Read comma-separated data from a character vector
}
\keyword{internal}
