% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_description_file}
\alias{read_description_file}
\title{Read dataset DESCRIPTION file}
\usage{
read_description_file(dataset_name)
}
\arguments{
\item{dataset_name}{Dataset name, character}
}
\value{
A \code{data.frame} with the following columns:
\item{Dataset}{Dataset name, character}
\item{Site_name}{Site name, character}
\item{Longitude}{Decimal Longitude, numeric}
\item{Latitude}{Decimal latitude, numeric}
\item{Instrument}{Instrument type, character}
\item{Ecosystem_type}{Ecosystem type, character}
\item{Primary_pub}{Primary publication, character}
\item{Other_pubs}{Other publications, character}
}
\description{
Read dataset DESCRIPTION file
}
\examples{
read_description_file("d20190406_TEST")
}
\keyword{internal}
