% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_ports_file}
\alias{read_ports_file}
\title{Read dataset PORTS file}
\usage{
read_ports_file(dataset_name, file_data = NULL)
}
\arguments{
\item{dataset_name}{Dataset name, character}

\item{file_data}{File data, character vector; optional for testing}
}
\value{
A \code{data.frame} with the following columns:
\item{CSR_PORT}{Port number, numeric; 0 = all ports}
\item{CSR_TREATMENT}{Treatment, character; by default "None"}
\item{CSR_SPECIES}{Species, character}
\item{CSR_DEPTH}{Depth of collar, cm}
\item{CSR_AREA}{Ground area of chamber, cm2}
}
\description{
Read dataset PORTS file
}
\keyword{internal}
