% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rearrange_columns}
\alias{rearrange_columns}
\title{Ensure required columns are present; put them first;
sort everything else alphabetically.}
\usage{
rearrange_columns(dsd, required_cols)
}
\arguments{
\item{dsd}{Dataset data, a data frame}

\item{required_cols}{Required column names, a character vector}
}
\value{
The column-sorted data frame.
}
\description{
Ensure required columns are present; put them first;
sort everything else alphabetically.
}
\note{
Called by \code{\link{read_raw_dataset}}.
}
\keyword{internal}
