% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_empty_columns}
\alias{remove_empty_columns}
\title{Remove empty (all-NA) columns from a data frame.}
\usage{
remove_empty_columns(x)
}
\arguments{
\item{x}{The data frame}
}
\value{
The data frame with all-empty columns removed.
}
\description{
Remove empty (all-NA) columns from a data frame.
}
\keyword{internal}
