% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_invalid_timestamps}
\alias{remove_invalid_timestamps}
\title{Remove rows with invalid timestamps.}
\usage{
remove_invalid_timestamps(dsd, tf, tz)
}
\arguments{
\item{dsd}{Dataset data, a \code{data.frame}}

\item{tf}{Timestamp format, character}

\item{tz}{Timezone, character}
}
\value{
The data frame with any NA-timestamp rows removed
}
\description{
Remove rows with invalid timestamps.
}
\keyword{internal}
