% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{run_all_reports}
\alias{run_all_reports}
\title{Run all individual dataset reports}
\usage{
run_all_reports(all_data, run_combined = TRUE)
}
\arguments{
\item{all_data}{A list of \code{cosore} datasets.}

\item{run_combined}{When done, run the combined (entire database) report?}
}
\value{
Nothing; run for its side effects.
}
\description{
Run all individual dataset reports
}
