% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{run_combined_report}
\alias{run_combined_report}
\title{Run a summary report on the entire COSORE database}
\usage{
run_combined_report(all_data, output_dir = "~/Desktop/Reports/")
}
\arguments{
\item{all_data}{A list of \code{cosore} datasets.}

\item{output_dir}{Output directory}
}
\value{
Nothing; run for its side effects.
}
\description{
Run a summary report on the entire COSORE database
}
