% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{run_reports}
\alias{run_reports}
\title{Run all individual dataset reports}
\usage{
run_reports(all_data)
}
\arguments{
\item{all_data}{A list of \code{cosore} datasets.}
}
\value{
Nothing; run for its side effects.
}
\description{
Run all individual dataset reports
}
