% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{run_single_report}
\alias{run_single_report}
\title{Run a single dataset report}
\usage{
run_single_report(ds)
}
\arguments{
\item{ds}{The dataset for which to generate a report.}
}
\value{
Nothing; run for its side effects.
}
\description{
Run a single dataset report
}
