% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_licor_file}
\alias{split_licor_file}
\title{Split a raw Licor (*.81x) file into pieces.}
\usage{
split_licor_file(filename, split_lines = 25000,
  out_dir = dirname(filename))
}
\arguments{
\item{filename}{Fully-qualified filename, character}

\item{split_lines}{Rough number of lines to split by, integer}

\item{out_dir}{Output directory, character}
}
\value{
Number of files created, invisibly.
}
\description{
Split a raw Licor (*.81x) file into pieces.
}
\note{
Licor output files can get \emph{very} big. This utility splits (on record
boundaries) a file into multiple sub-files.
}
\keyword{internal}
