% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{substitute_release_info}
\alias{substitute_release_info}
\title{Substitute Git SHA, date, etc., into file data.}
\usage{
substitute_release_info(f_data, file_descriptions, db_size, force)
}
\arguments{
\item{f_data}{File data, a character vector}

\item{file_descriptions}{A character vector of filenames and their descriptions}

\item{db_size}{Database size, numeric}

\item{force}{Ignore git dirty status? Logical}
}
\value{
The file data with substitutions made.
}
\description{
Substitute Git SHA, date, etc., into file data.
}
\note{
Called only by \code{\link{csr_make_release}}.
}
\keyword{internal}
