% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{TSM_change}
\alias{TSM_change}
\title{Change "T2", "SM4.5", etc., to "Tx" and "SMx"}
\usage{
TSM_change(x, prefixes = c("CSR_T", "CSR_SM"))
}
\arguments{
\item{x}{Vector of dataset names}

\item{prefixes}{Prefixes}
}
\value{
A transformed vector of names.
}
\description{
Change "T2", "SM4.5", etc., to "Tx" and "SMx"
}
\keyword{internal}
