% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{write_dataset_data}
\alias{write_dataset_data}
\title{Substitute Git SHA, date, etc., into file data.}
\usage{
write_dataset_data(dataset_name, ds, path, db_size)
}
\arguments{
\item{dataset_name}{Dataset name, character}

\item{ds}{Dataset, a list}

\item{path}{Output path, character}

\item{db_size}{Database size, numeric}
}
\value{
Updated database size.
}
\description{
Substitute Git SHA, date, etc., into file data.
}
\note{
Called only by \code{\link{csr_make_release}}.
}
\keyword{internal}
