% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_stan_data}
\alias{write_stan_data}
\title{Write standardized data, diagnostics, and perhaps a gitignore file}
\usage{
write_stan_data(dataset, outpath)
}
\arguments{
\item{dataset}{Dataset, a list}

\item{outpath}{Output path to write to, character}
}
\value{
Nothing.
}
\description{
Write standardized data, diagnostics, and perhaps a gitignore file
}
\note{
This is called only by \code{\link{csr_standardize_data}}.
}
\keyword{internal}
