% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{combine_data}
\alias{combine_data}
\title{Title}
\usage{
combine_data(datasets, ...)
}
\arguments{
\item{datasets}{Character vector of dataset names}

\item{...}{Dataset objects}
}
\value{
x
}
\description{
Title
}
