% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{csr_build}
\alias{csr_build}
\title{Build the COSORE database}
\usage{
csr_build(
  raw_data,
  dataset_names = list_datasets(),
  force_raw = FALSE,
  write_standardized = FALSE,
  standardized_path = "./inst/extdata/datasets",
  quiet = FALSE
)
}
\arguments{
\item{raw_data}{The raw data folder to use, character path}

\item{dataset_names}{The raw data folder to use, character path}

\item{force_raw}{Always read raw (as opposed to standardized) data? Logical}

\item{write_standardized}{Write standardized data after parsing? Logical}

\item{standardized_path}{Output path (typically \code{inst/extdata/datasets})
for standardized data, character}

\item{quiet}{Print progress messages and warnings? Logical}
}
\value{
All the built data, invisibly.
}
\description{
Build the COSORE database
}
