% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{csr_database}
\alias{csr_database}
\title{Return metadata for the entire COSORE database.}
\usage{
csr_database()
}
\value{
A \code{data.frame} with metadata about each constituent dataset. This consists of parts of
the \code{description} file, joined with parts of \code{ports} and \code{diag}.
}
\description{
Return metadata for the entire COSORE database.
}
