% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{csr_database}
\alias{csr_database}
\title{Return metadata (the \code{description} file) for the entire COSORE database.}
\usage{
csr_database()
}
\value{
A \code{data.frame} with metadata about each constituent dataset.
}
\description{
Return metadata (the \code{description} file) for the entire COSORE database.
}
