% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{csr_dataset}
\alias{csr_dataset}
\title{Return a single COSORE dataset.}
\usage{
csr_dataset(dataset_name, quiet = FALSE, metadata_only = FALSE)
}
\arguments{
\item{dataset_name}{The dataset name (from \code{\link{list_datasets}}), character}

\item{quiet}{Print progress messages and warnings? Logical}

\item{metadata_only}{Quick-read metadata only? Logical}
}
\value{
A list with (at least) elements:
\item{description}{Contents of \code{DESCRIPTION.txt} file. This contains site,
instrument, and publication information.}
\item{contributors}{Contents of \code{CONTRIBUTORS.txt} file with dataset contributor
information. The first contributor listed is assumed to the main contact.}
\item{ports}{Contents of \code{PORTS.txt} file, describing the species and measurement
conditions of different ports (typically multiple chambers are measured by a single
gas analyzer).}
\item{data}{Continuous soil respiration data, parsed into a \code{data.frame}.}
\item{diagnostics}{Diagnostics on the data parsing and QC process.}
\item{ancillary}{Ancillary site information.}
}
\description{
Return a single COSORE dataset.
}
