% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{csr_make_release}
\alias{csr_make_release}
\title{Make a new COSORE release}
\usage{
csr_make_release(all_data, path, vignette_rebuilt = FALSE,
  force = FALSE)
}
\arguments{
\item{all_data}{A list of \code{cosore} datasets}

\item{path}{Path to write files to; must already exist}

\item{vignette_rebuilt}{Has vignette been rebuilt? Logical}

\item{force}{Ignore git dirty status? Logical}
}
\value{
Fully qualified name of zip file containing release.
}
\description{
Make a new COSORE release
}
\details{
To make a new release, the git working directory
must be clean (unless \code{force = TRUE}), a \code{path} must
be specified, and there should be file descriptors (in the code)
for all files that will be zipped into the release;
this information is inserted into the \code{README.md} file.
Dynamic information like the current date, git commit hash,
database version, and
database size are all copied into \code{README.md} as well.
}
