% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{csr_standardize_data}
\alias{csr_standardize_data}
\title{Write out standardized data and diagnostics tables}
\usage{
csr_standardize_data(all_data, path, create_dirs = FALSE)
}
\arguments{
\item{all_data}{A list of \code{cosore} datasets}

\item{path}{Output path, character}

\item{create_dirs}{Create subdirectories as needed? Logical}
}
\value{
Nothing.
}
\description{
This is typically used to write standardized data
to \code{inst/extdata/datasets}. These data can be removed
by \code{\link{csr_remove_stan_data}}.
}
