% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user.R
\name{csr_table}
\alias{csr_table}
\title{Extract one table (\code{description}, \code{data}, etc.) of all datasets}
\usage{
csr_table(all, table)
}
\arguments{
\item{all}{All data, a list of lists}

\item{table}{Name of table to extract, character}
}
\value{
A \code{\link{data.frame}}.
}
\description{
Extract one table (\code{description}, \code{data}, etc.) of all datasets
}
