% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{extract_line}
\alias{extract_line}
\title{Parse labels from information files}
\usage{
extract_line(
  file_data,
  line_label,
  required = TRUE,
  sep = ":",
  numeric_data = FALSE
)
}
\arguments{
\item{file_data}{Character vector of file lines}

\item{line_label}{Character label}

\item{sep}{Separator}

\item{numeric_data}{Numeric data? Logical}
}
\value{
Value (character) of data opposite label
}
\description{
Parse labels from information files
}
\note{
If label is not found, produces an error (if label was required) or returns
\code{NA} (if not).
}
\keyword{internal}
