% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{list_datasets}
\alias{list_datasets}
\title{List datasets available to build}
\usage{
list_datasets(path = resolve_dataset(""))
}
\arguments{
\item{path}{Path name}
}
\value{
Character vector of datasets.
}
\description{
List datasets available to build
}
\details{
This returns a vector of metadata folder names available in
the package's \code{inst/exdata} folder. The \code{\link{csr_build}}
function uses these names, in combination with a user-specified raw
data folder, to build the database.
}
\examples{
list_datasets()
}
