% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-LI8100.R
\name{parse_LI-8100A_RAW}
\alias{parse_LI-8100A_RAW}
\title{Loop through directory and read raw multiplexed Licor-8100 data}
\usage{
"parse_LI-8100A_RAW"(path, UTC_offset)
}
\arguments{
\item{path}{Data directory path, character}

\item{UTC_offset}{Offset from UTC in hours, numeric}
}
\value{
A data frame with all data read from file(s).
}
\description{
Loop through directory and read raw multiplexed Licor-8100 data
}
