% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-LI8100.R
\name{parse_LI8100_raw}
\alias{parse_LI8100_raw}
\title{Loop through directory and read raw multiplexed Licor-8100 data}
\usage{
parse_LI8100_raw(path, port_data, UTC_offset)
}
\arguments{
\item{path}{Directory path, character}

\item{port_data}{Port data, a list returned by \code{\link{read_ports_file}}}

\item{UTC_offset}{Offset from UTC in hour, numeric}
}
\value{
A data frame with all data read from file(s).
}
\description{
Loop through directory and read raw multiplexed Licor-8100 data
}
