% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qaqc.R
\name{qaqc_data}
\alias{qaqc_data}
\title{QA/QC a dataset.}
\usage{
qaqc_data(dsd, diag, remove_na = TRUE, remove_error = TRUE,
  remove_flux = TRUE, remove_temp = TRUE, flux_limits = c(-10, 50),
  temp_limits = c(-50, 60))
}
\arguments{
\item{dsd}{Dataset data, a tibble}

\item{diag}{Diagnostics data, a tibble}

\item{remove_na}{Remove NA flux values? Logical}

\item{remove_error}{Remove records marked as errors? Logical}

\item{remove_flux}{Remove extreme flux values? Logical}

\item{remove_temp}{Remove extreme temperature values? Logical}

\item{flux_limits}{Two-value numeric vector giving flux limits}

\item{temp_limits}{Two-value numeric vector giving temperature limits}
}
\value{
A list with the `dsd` and `diag` tibbles.
}
\description{
QA/QC a dataset.
}
\note{
This is normally called from \code{\link{read_raw_dataset}}.
}
