% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_contributors_file}
\alias{read_contributors_file}
\title{Read dataset CONTRIBUTORS file}
\usage{
read_contributors_file(dataset_name, file_data = NULL)
}
\arguments{
\item{dataset_name}{Dataset name, character}

\item{file_data}{File data, character vector; optional for testing}
}
\value{
A \code{data.frame} with the following columns:
\item{CSR_FIRST_NAME}{Site name, character}
\item{CSR_FAMILY_NAME}{Decimal Longitude, numeric}
\item{CSR_EMAIL}{Decimal latitude, numeric}
\item{CSR_ORCID}{Instrument type, character}
\item{CSR_ROLE}{Role, character}
}
\description{
Read dataset CONTRIBUTORS file
}
\note{
For information about ORCID see \url{https://orcid.org}. For
CRediT roles, see \url{https://www.casrai.org/credit.html}.
}
\keyword{internal}
