% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_file}
\alias{read_file}
\title{Read a text file and strip out comments}
\usage{
read_file(dataset_name, file_name, file_data = NULL,
  comment_char = "#")
}
\arguments{
\item{dataset_name}{Dataset name, character}

\item{file_name}{File name, character}

\item{file_data}{File data, character vector; optional for testing}

\item{comment_char}{Start-of-line comment character}
}
\value{
Contents of file in a character vector.
}
\description{
Read a text file and strip out comments
}
\keyword{internal}
