% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{read_ports_file}
\alias{read_ports_file}
\title{Read dataset PORTS file}
\usage{
read_ports_file(dataset_name)
}
\arguments{
\item{dataset_name}{Dataset name, character}
}
\value{
A \code{data.frame} with the following columns:
\item{Dataset}{Dataset name, character}
\item{Port}{Port number, numeric; 0 = all ports}
\item{Treatment}{Treatment, character; by default "None"}
\item{Species}{Species, character}
\item{V1-V4}{Sensor information for options V1-V4 probes}
}
\description{
Read dataset PORTS file
}
\examples{
read_contributors_file("d20190406_TEST")
}
\keyword{internal}
