% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{resolve_dataset}
\alias{resolve_dataset}
\title{Get the full path of a dataset folder}
\usage{
resolve_dataset(dataset_name)
}
\arguments{
\item{dataset_name}{Dataset name, character}
}
\value{
Fully-qualified filename of dataset folder in \code{inst/extdata}.
}
\description{
Get the full path of a dataset folder
}
\examples{
resolve_dataset("d20190406_TEST")
}
\keyword{internal}
