% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{run_single_report}
\alias{run_single_report}
\title{Run a single dataset report}
\usage{
run_single_report(ds, output_dir = "~/Desktop/Reports/", quiet = FALSE,
  quick = FALSE)
}
\arguments{
\item{ds}{The dataset for which to generate a report.}

\item{output_dir}{Output directory, character}

\item{quiet}{Passed on to \code{rmarkdown::render}}

\item{quick}{Passed on to \code{rmarkdown::render}}
}
\value{
Nothing; run for its side effects.
}
\description{
Run a single dataset report, writing the resulting \code{html}
file into a specified directory.
}
